fix: Move activity and market event to types
