+++ src/views/Nft/market/ActivityHistory/ActivityHistory.tsx
@@ -16,7 +16,7 @@ import {
 import { getCollectionActivity, getNftsFromDifferentCollectionsApi } from 'state/nftMarket/helpers'
 import Container from 'components/Layout/Container'
 import TableLoader from 'components/TableLoader'
-import { Collection, NftToken, TokenIdWithCollectionAddress } from 'state/nftMarket/types'
+import { Activity, Collection, NftToken, TokenIdWithCollectionAddress } from 'state/nftMarket/types'
 import { useTranslation } from 'contexts/Localization'
 import { useBNBBusdPrice } from 'hooks/useBUSDPrice'
 import useTheme from 'hooks/useTheme'
@@ -25,7 +25,6 @@ import { useGetNftActivityFilters } from 'state/nftMarket/hooks'
 import { Arrow, PageButtons } from '../components/PaginationButtons'
 import NoNftsImage from '../components/Activity/NoNftsImage'
 import ActivityFilters from './ActivityFilters'
-import { Activity } from '../types/Activity'
 import ActivityRow from '../components/Activity/ActivityRow'
 import { sortActivity } from './utils/sortActivity'
