+++ src/views/Nft/market/Collection/IndividualNFTPage/OneOfAKindNftPage/ActivityCard.tsx
@@ -2,10 +2,9 @@ import React, { useEffect, useState } from 'react'
 import { Flex, Card, Text, useMatchBreakpoints, Table, Th, ArrowBackIcon, ArrowForwardIcon } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import useTheme from 'hooks/useTheme'
-import { NftToken } from 'state/nftMarket/types'
+import { Activity, NftToken } from 'state/nftMarket/types'
 import { useBNBBusdPrice } from 'hooks/useBUSDPrice'
 import { useAppDispatch } from '../../../../../../state'
-import { Activity } from '../../../types/Activity'
 import NoNftsImage from '../../../components/Activity/NoNftsImage'
 import TableLoader from '../../../../../../components/TableLoader'
 import { Arrow, PageButtons } from '../../../components/PaginationButtons'
