+++ src/views/Nft/market/Profile/components/ActivityHistory/index.tsx
@@ -7,7 +7,7 @@ import { useAppDispatch } from 'state'
 import { useUserNfts } from 'state/nftMarket/hooks'
 import { ArrowBackIcon, ArrowForwardIcon, Card, Flex, Table, Text, Th, useMatchBreakpoints } from '@pancakeswap/uikit'
 import { getNftsFromDifferentCollectionsApi, getUserActivity } from 'state/nftMarket/helpers'
-import { NftToken, TokenIdWithCollectionAddress, UserNftInitializationState } from 'state/nftMarket/types'
+import { Activity, NftToken, TokenIdWithCollectionAddress, UserNftInitializationState } from 'state/nftMarket/types'
 import { useTranslation } from 'contexts/Localization'
 import TableLoader from 'components/TableLoader'
 import { useBNBBusdPrice } from 'hooks/useBUSDPrice'
@@ -16,7 +16,6 @@ import { useParams } from 'react-router'
 import { sortUserActivity } from '../../utils/sortUserActivity'
 import NoNftsImage from '../../../components/Activity/NoNftsImage'
 import { Arrow, PageButtons } from '../../../components/PaginationButtons'
-import { Activity } from '../../../types/Activity'
 import ActivityRow from '../../../components/Activity/ActivityRow'
 
 const MAX_PER_PAGE = 8
