+++ src/views/Nft/market/components/Activity/ActivityEventText.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Text, TextProps } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
-import { MarketEvent } from '../../types/MarketEvent'
+import { MarketEvent } from '../../../../../state/nftMarket/types'
 
 interface ActivityEventTextProps extends TextProps {
   marketEvent: MarketEvent
