--- src/views/Nft/market/types/MarketEvent.tsx
@@ -1,7 +1,0 @@
-export enum MarketEvent {
-  NEW = 'NEW',
-  CANCEL = 'CANCEL',
-  MODIFY = 'MODIFY',
-  BUY = 'BUY',
-  SELL = 'SELL',
-}
